[REFACTOR] unused library
