+++ src/main/java/shoppingmall/core/domain/Goods/Goods.java
@@ -1,6 +1,5 @@
 package shoppingmall.core.domain.Goods;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
