+++ src/main/java/shoppingmall/core/web/controller/MemberController.java
@@ -4,11 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import shoppingmall.core.config.JwtTokenProvider;
 import shoppingmall.core.domain.member.Member;
-import shoppingmall.core.domain.member.MemberRepository;
 import shoppingmall.core.service.member.MemberService;
 import shoppingmall.core.web.dto.*;
 import shoppingmall.core.web.dto.member.MemberCreateRequestDto;
