+++ src/main/java/shoppingmall/core/web/dto/goods/GoodsCreateRequestDto.java
@@ -4,11 +4,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import shoppingmall.core.domain.Goods.Goods;
-
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.validation.constraints.NotEmpty;
 
 @Getter
